// src/lib/api/generated/models/remoteAuthConfig.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.1
 */

export type RemoteAuthConfig = {
  type: 'none';
} | {
  secret: string;
  type: 'bearer_secret';
} | {
  secret: string;
  type: 'basic_secret';
  username: string;
} | {
  header: string;
  secret: string;
  type: 'api_key_secret';
};